feat: be able to auth

+++ src/encrypt/SrunBase64.cpp
@@ -9,14 +9,14 @@ static std::string alpha = "LVoJPiCN2R8G90yg+hmFHuacZ1OWMnrsSTXkYpUq/3dlbfKwv6xz
 
 std::string get_base64_string(std::string s) {
     int i;
-    int b10;
+    uint32_t b10;
     std::vector<BYTE> x;
     int imax = s.length() - s.length() % 3;
     if (s.length() == 0) {
         return s;
     }
     for (i = 0; i < imax; i += 3) {
-        b10 = (s[i] << 16) | (s[i + 1] << 8) | s[i + 2];
+        b10 = (((BYTE)s[i]) << 16) | (((BYTE)s[i+1]) << 8) | ((BYTE)s[i+2]);
         x.push_back(alpha[(b10 >> 18)]);
         x.push_back(alpha[((b10 >> 12) & 63)]);
         x.push_back(alpha[((b10 >> 6) & 63)]);
@@ -25,14 +25,14 @@ std::string get_base64_string(std::string s) {
     i = imax;
     switch (s.length() - imax) {
         case 1:
-            b10 = s[i] << 16;
+            b10 = ((BYTE)s[i]) << 16;
             x.push_back(alpha[(b10 >> 18)]);
             x.push_back(alpha[((b10 >> 12) & 63)]);
             x.push_back(pad_char);
             x.push_back(pad_char);
             break;
         case 2:
-            b10 = (s[i] << 16) | (s[i + 1] << 8);
+            b10 = (((BYTE)s[i]) << 16) | (((BYTE)s[i+1]) << 8);
             x.push_back(alpha[(b10 >> 18)]);
             x.push_back(alpha[((b10 >> 12) & 63)]);
             x.push_back(alpha[((b10 >> 6) & 63)]);
